Adding tesseract OCR
