+++ client/src/App.js
@@ -1,9 +1,8 @@
 import React, { Fragment } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Landing from './views/Landing';
-import Login from './components/Auth/Login';
+import Login from './components/auth/Login';
 import Dashboard from './views/Dashboard/Dashboard';
-import Home from './views/SidebarPages/Home';
 
 const App = () => (
   <Router>
